requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.1
camelcase==0.2
click==8.0.1
colorama==0.4.4
distlib==0.3.4
Django==3.2.7
django-cors-headers==3.9.0
Faker==11.3.0
filelock==3.4.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.26
mysqlclient==2.0.3
platformdirs==2.4.0
pycodestyle==2.7.0
PyJWT==1.4.2
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.26
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
virtualenv==20.10.0
Werkzeug==2.2.2
zipp==3.8.0
